Main

Description

The purpose of this application is to Download new data from iTunes URL and sotre it in a SQLite databse. Main Functionality is as Follows: - The user has an option to download using Artist Name and/or Genre Type - Additional features are to list all the Genres and Artist - As well as search and list all the songs by a particular Genre - Also, the user can perform operations like Add, Delete and View a Playlist

Written by Shashank Singh, Naman Shrivastava, Vishal and Ziyad Sheikh

Documentation

main :: IO ()

Options below are for adding the details to the database. | A lots of exceptions are being handled here!